Fix two warnings in bench timers. http://codereview.appspot.com/5164049/
git-svn-id: http://skia.googlecode.com/svn/trunk@2400 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ bench/BenchSysTimer_windows.cpp
@@ -42,7 +42,7 @@ void BenchSysTimer::startCpu() {
 
 double BenchSysTimer::endCpu() {
     ULONGLONG end_cpu = winCpuTime();
-    return (end_cpu - this->fStartCpu) / 10000;
+    return static_cast<double>((end_cpu - this->fStartCpu)) / 10000.0L;
 }
 double BenchSysTimer::endWall() {
     LARGE_INTEGER end_wall;
@@ -55,8 +55,9 @@ double BenchSysTimer::endWall() {
     
     LARGE_INTEGER frequency;
     if (0 == ::QueryPerformanceFrequency(&frequency)) {
-        return 0;
+        return 0.0L;
     } else {
-        return (double)ticks_elapsed.QuadPart / frequency.QuadPart * 1000;
+        return static_cast<double>(ticks_elapsed.QuadPart)
+            /  static_cast<double>(frequency.QuadPart * 1000);
     }
 }
